WEBAPP-208: Fix moving of files
